refactor(config): 拆分 config.go — scanner 子系统类型移入 scanner_config.go (#160)
config.go (682 行 / 33KB god-file) 拆分：scanner 子系统的 ~230 行配置类型
(ScannerConfig + RouterARPConfig + RDNSConfig + MDNSConfig + ARPScanConfig +
EBPFConfig + DiscoveryConfig + DiscoverySourceToggle + HostapdDiscoveryConfig +
DNSLogDiscoveryConfig + PipelineDefaultsConfig) 移入 scanner_config.go。

config.go 保留：Config 根结构 + 简单类型 (Server/DB/Auth/Log/CORS/...) +
Load() + Validate() + normalize* 函数。纯代码移动，无逻辑变更。

config.go 682→455 行 (-33%)，scanner_config.go 244 行。
